refactor(runner-pi): move pi stream parsing out of platform into runner-pi
parsePiStreamLine + processPiLogs are Pi-specific stdout → RunEvent
adaptation logic, not platform concerns. Any consumer of the Pi runtime
(CLI, GitHub Action, remote runners) reuses the same coalescing +
shape validation, so they belong next to the Pi runner.

- packages/runner-pi: new stream-parser.ts + test/stream-parser.test.ts
  (ports the full coalescing + breadcrumb-wrap coverage)
- apps/api: declare @appstrate/runner-pi workspace dep, import from it
- apps/api/pi.ts: strip the 120 LOC of local parser + log-processor
- apps/api tests: pi-parser.test.ts deleted; pi-adapter.test.ts reduced
  to deriveKeyPlaceholder (the one platform-only helper left)
