+++ apps/api/package.json
@@ -20,6 +20,7 @@
     "@appstrate/db": "workspace:*",
     "@appstrate/emails": "workspace:*",
     "@appstrate/env": "workspace:*",
+    "@appstrate/runner-pi": "workspace:*",
     "@appstrate/shared-types": "workspace:*",
     "@better-auth/oauth-provider": "^1.6.5",
     "@hono/swagger-ui": "^0.6.1",
